// src/app/generated/blueprint.api/model/catalogInject.ts
/**
 * Blueprint API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Inject } from './inject';


export interface CatalogInject { 
    id?: string;
    catalogId?: string;
    injectId?: string;
    inject?: Inject;
}